moved file-upload route implementation into the service layer
